Small improvements for QoL
A few small improvements for quality of life; might be worth
upstreaming as I was looking into this plugin.

1. create a friendly Nix wrapper that includes maven and some sane
defaults such as setting the maven repository to a temporary directory.

This causes the build to take longer since nothing is cached,
unfortunately however the likelyhood of using some local dependency is
less.

2. change the default to emit to STDOUT just to follow unix philosophy a
bit more. I changed the logging that maven emits to STDERR instead in
similar vein.

3. updated the README accordingly